Don't allow markup in key or desc config
Before the key and desc values where added with pango's
set_markup function. This fails when the key is "<" as pango
expects this to be the start of a markup tag.

By using set_text instead this problem is solved with the
drawback of not allowing markup in the description of a key.
However the documentation does not advertise that the
description could contain markup nor does it seem to be
a sensible feature.

Fixes: https://github.com/MaxVerevkin/wlr-which-key/issues/26